Fix order and more tag bloger

+++ AgriculturalWarehouseManagement/src/main/java/com/example/AgriculturalWarehouseManagement/Backend/services/bloger/BlogCategoryService.java
@@ -1,7 +1,7 @@
 package com.example.AgriculturalWarehouseManagement.Backend.services.bloger;
 
 import com.example.AgriculturalWarehouseManagement.Backend.models.BlogCategory;
-import com.example.AgriculturalWarehouseManagement.Backend.repositorys.blog.BlogCategoryRepository;
+import com.example.AgriculturalWarehouseManagement.Backend.repositorys.bloger.BlogCategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,7 +16,7 @@ public class BlogCategoryService implements IBlogCategoryService {
     private final BlogCategoryRepository blogCategoryRepository;
 
     @Override
-    public BlogCategory findById(Long id) {
+    public BlogCategory findById(Integer id) {
         return blogCategoryRepository.findById(id).orElse(null);
     }
